requirements.txt
# Streamlit Community Cloud installs from this file (Python 3.9+ recommended).
streamlit>=1.28.0,<2
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
nltk>=3.6
joblib>=1.0.0
plotly>=5.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.2
langgraph>=0.2.0
groq>=0.4.0
chromadb>=0.5.0
google-generativeai>=0.8.0
python-dotenv>=0.19.0
